Updated UI for Login and Registration
