+++ src/pages/Landing.js
@@ -1,9 +1,10 @@
 import React from "react";
+import {headerStyle, contentStyle, footerStyle, formStyle, inputStyle } from '../stylesheets/mainUI';
 
 function Landing() {
   return (
     <div>
-      
+      landing
     </div>
   );
 }
